Refactor package structure for improved organization: - Moved locking-related classes to `system.notification.email.locking`. - Relocated security and recovery-related tests to appropriate packages. - Updated imports across affected files.
