+++ src/main/java/info/mackiewicz/bankapp/system/notification/email/locking/AccountLockingStrategy.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.system.locking;
+package info.mackiewicz.bankapp.system.notification.email.locking;
 
 import info.mackiewicz.bankapp.account.exception.AccountLockException;
 import lombok.Getter;
