+++ src/main/java/info/mackiewicz/bankapp/system/notification/email/locking/LockingConfig.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.system.locking;
+package info.mackiewicz.bankapp.system.notification.email.locking;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
