+++ src/test/java/info/mackiewicz/bankapp/system/token/service/TokenOperationsServiceTest.java
@@ -1,6 +1,5 @@
-package info.mackiewicz.bankapp.system.security.token.service;
+package info.mackiewicz.bankapp.system.token.service;
 
-import info.mackiewicz.bankapp.system.token.service.TokenOperationsService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
